--- electron-preload/index.ts
@@ -1,19 +1,0 @@
-import { useLoading } from './loading'
-
-function domReady(condition: DocumentReadyState[] = ['complete', 'interactive']) {
-  return new Promise(resolve => {
-    if (condition.includes(document.readyState)) {
-      resolve(true)
-    } else {
-      document.addEventListener('readystatechange', () => {
-        if (condition.includes(document.readyState)) {
-          resolve(true)
-        }
-      })
-    }
-  })
-}
-
-const { appendLoading, removeLoading } = useLoading()
-window.removeLoading = removeLoading
-domReady().then(appendLoading)
